requirements.txt
beautifulsoup4==4.11.1
dnspython==2.2.1
Flask==2.0.2
Flask_Cors==3.0.10
pymongo==4.2.0
python-dotenv==0.20.0
gunicorn==20.1.0
selenium==4.3.0